+++ lib/http_instrumentation.rb
@@ -60,18 +60,20 @@ module HTTPInstrumentation
     # block. If no block is given, then return the current HTTP client name.
     #
     # @param name [String, Symbol, nil] The name of the client to set
-    # @return [String, Symbol, nil] The current name of the client
+    # @return [String, Symbol, nil] If a block is given, then the return value of the block. Otherwise
+    #   the current client name.
     def client(name = nil)
       save_val = Thread.current[:http_instrumentation_client]
       if block_given?
         begin
-          Thread.current[:http_instrumentation_client] = name
+          Thread.current[:http_instrumentation_client] = name&.to_s
           yield
         ensure
           Thread.current[:http_instrumentation_client] = save_val
         end
+      else
+        save_val
       end
-      save_val
     end
 
     # Returns true if instrumentation is currently silenced.